requirements.txt
flask
flask-cors
agno
google-genai
python-dotenv 